requirements.txt
--index-url https://download.pytorch.org/whl/cu124
--extra-index-url https://pypi.org/simple

torch
torchaudio

gradio
chatterbox-tts
numpy
librosa
noisereduce
soundfile
